+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The `Unreleased` section name is replaced by the expected version of next releas
 
 ## [Unreleased]
 
+### Fixed
+
+- `Tool`: Corrected `NullReferenceException` when using `StreamLike` option
+
 <a name="4.1.0"></a>
 ## [4.1.0] - 2025-06-12